+++ CHANGELOG.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### 17 - Refactor `FireManager` - 2020-08-23
+
+#### ADDED
+
+- New `FireManager` interface has a void `FireManager#startFires()` method
+- `DefaultFireManager` updated to implement `FireManager`
+- `FlammabilityChecker` and default implementation `DefaultFlammabilityChecker`
+  that indicate whether a specific grid cell is flammable or not
+
+#### CHANGED
+
+- `FireManager` is now an interface and the old class of the same name is now
+  called `DefaultFireManager`
+- `DefaultFireManager#numFires()` now package protected
+- `DefaultFireManager` now accepts a `FlammabilityChecker` in its constructor
+
 ### 16 - Replace land-cover colonisation model with CompletelySpatiallyRandomColoniser - 2020-08-13
 
 We decided to use a simpler implementation of spatially random seed